requirements.txt
pyqt6-sip~=13.10
PyQt6~=6.10
PyQt6-WebEngine~=6.0
requests~=2.0
python-dotenv~=1.0
requests~=2.0
frozendict~=2.4